feat(BUX-727): tx direction filter - removed
